docs(generate): Adiciona novos arquivos de documentação e experimenta diferentes formas de criação
Adiciona novos arquivos de documentação e experimenta diferentes formas de criação de documentação para a API, com o objetivo de encontrar o melhor processo para o projeto. A documentação é fundamental para garantir a compreensão do código por outras pessoas e facilitar a manutenção e a escalabilidade do projeto. A adição de novos arquivos de documentação e a experimentação com diferentes métodos de criação podem melhorar a colaboração entre os membros da equipe e o uso do projeto por terceiros.
