+++ .pre-commit-config.yaml
@@ -37,3 +37,14 @@ repos:
         types: [python]
         pass_filenames: false
         always_run: true
+
+  - repo: local
+    hooks:
+      - id: docs
+        name: docs
+        stages: [commit]
+        language: system
+        entry: python -m docs.generate_documentation
+        types: [python]
+        pass_filenames: false
+        always_run: true
