+++ docs/index.html
@@ -4,22 +4,70 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>API Documentation</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
-      href="https://fonts.googleapis.com/css?family=Montserrat:300,400,700|Roboto:300,400,700"
+      href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap"
       rel="stylesheet"
     />
     <style>
       body {
-        margin: 0;
-        padding: 0;
+        font-family: "Roboto", sans-serif;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 100vh;
+        background-color: #f4f4f4;
+      }
+      .card {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        flex-direction: column;
+        background-color: #fff;
+        border-radius: 8px;
+        box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+        padding: 32px;
+        margin: 0 16px;
+        text-align: center;
+        width: 300px;
+        max-width: 90%;
+      }
+      .card h2 {
+        font-weight: 500;
+        font-size: 24px;
+        margin-bottom: 16px;
+      }
+      .card p {
+        font-weight: 400;
+        font-size: 16px;
+        margin-bottom: 32px;
+      }
+      .card a {
+        background-color: #232ca599;
+        color: #fff;
+        padding: 8px 16px;
+        border-radius: 4px;
+        text-decoration: none;
+        font-weight: 500;
+        font-size: 16px;
+        transition: background-color 0.2s ease-in-out;
+      }
+      .card a:hover {
+        background-color: #232ca5;
       }
     </style>
-    <script src="https://cdn.jsdelivr.net/npm/redoc@next/bundles/redoc.standalone.js"></script>
   </head>
   <body>
-    <redoc spec-url="openapi.json"></redoc>
-    <script>
-      Redoc.init("openapi.json", {}, document.querySelector("redoc"));
-    </script>
+    <div class="card">
+      <h2>Redoc</h2>
+      <p>Visualize a documentação da API usando o formato Redoc.</p>
+      <a href="redoc.html">View Documentation</a>
+    </div>
+    <div class="card">
+      <h2>Swagger UI</h2>
+      <p>Visualize a documentação da API usando o formato Swagger UI.</p>
+      <a href="swagger.html">View Documentation</a>
+    </div>
   </body>
 </html>
